Optimizing code

+++ components/home/ToDoListItem.jsx
@@ -1,45 +1,52 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import styles from "./ToDoListItem.module.css";
 import Modal from "../ui/Modal";
 import { RiDeleteBin6Line } from "react-icons/ri";
 import { useRouter } from "next/router";
 
-const ToDoListItem = (props) => {
-  const { task, username } = props;
+const ToDoListItem = ({ task, username }) => {
   const [isLoading, setIsLoading] = useState(false);
+  const [isError, setIsError] = useState(false);
+  const [errorMessage, setErrorMessage] = useState("");
   const [isDone, setIsDone] = useState(task.status);
   const router = useRouter();
 
-  const checkboxHandler = async (event) => {
+  const changeHandler = async (task, operation) => {
     setIsLoading(true);
-    const task = event.target.innerText;
-    const status = !isDone;
+    let response;
 
-    const response = await fetch("/api/task/updateTask", {
-      method: "PATCH",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ task, status, username: props.username }),
-    });
+    if (operation === "Update") {
+      const status = !isDone;
 
-    if (!response.ok) {
-      console.log("failed");
+      response = await fetch("/api/task/updateTask", {
+        method: "PATCH",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ task, status, username }),
+      });
+
+      if (response.ok) {
+        setIsDone(status);
+      }
+      setIsLoading(false);
       return;
     }
 
-    setIsDone(status);
-    setIsLoading(false);
-  };
-
-  const deleteHandler = async (task) => {
-    setIsLoading(true);
-    const response = await fetch("/api/task/deleteTask", {
-      method: "DELETE",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ task, username }),
-    });
+    if (operation === "Delete") {
+      response = await fetch("/api/task/updateTask", {
+        method: "DELETE",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ task, username }),
+      });
+    }
 
     if (!response.ok) {
-      console.log("failed");
+      const data = await response.json();
+      setErrorMessage(data.message);
+      setIsLoading(false);
+      setIsError(true);
+      setTimeout(() => {
+        setIsError(false);
+      }, 1500);
       return;
     }
 
@@ -53,19 +60,25 @@ const ToDoListItem = (props) => {
 
   return (
     <>
+      {isError && <Modal message={errorMessage} />}
       {isLoading && <Modal message={"Updating task..."} />}
       <li className={styles.listItem}>
-        <div className={styles.taskContainer} onClick={checkboxHandler}>
+        <div
+          className={styles.taskContainer}
+          onClick={() => changeHandler(task.task, "Update")}
+        >
           <input
             type="checkbox"
-            onChange={checkboxHandler}
-            onClick={checkboxHandler}
+            onChange={changeHandler}
+            onClick={changeHandler}
             checked={isDone ? true : false}
           ></input>
           <span style={textStyle}>{task.task}</span>
         </div>
         <div className={styles.deleteIcon}>
-          <RiDeleteBin6Line onClick={() => deleteHandler(task.task)} />
+          <RiDeleteBin6Line
+            onClick={() => changeHandler(task.task, "Delete")}
+          />
         </div>
       </li>
     </>
